refactor(desktop): extract inline test modules into src/tests/
Move the four remaining inline #[cfg(test)] mod tests blocks out of
production source files and into dedicated files under src/tests/,
matching the pattern already used by auth/api/songs/filesystem.

- error.rs  -> tests/error_tests.rs
- lib.rs    -> tests/lib_tests.rs   (include_bytes! path adjusted:
                                      ../icons -> ../../icons)
- updater.rs -> tests/updater_tests.rs
- models.rs -> tests/models_tests.rs

Each source file now ends with the same 3-line declaration:

    #[cfg(test)]
    #[path = "tests/{module}_tests.rs"]
    mod tests;

No test logic changed — pure file move. 281 tests pass, clippy clean.
